add auth to requests

+++ frontend/src/components/WorkoutDetails.js
@@ -1,19 +1,28 @@
+import { useAuthContext } from '../hooks/useAuthContext'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
+  const { user } = useAuthContext()
   const { dispatch } = useWorkoutsContext()
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
+
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
     if (response.ok) {
-      dispatch({type: 'DELETE_WORKOUT', payload: json})
+      dispatch({ type: 'DELETE_WORKOUT', payload: json })
     }
   }
